Crate namada_wallet

Provides functionality for managing keys and addresses for a user

Re-exports

pub use pre_genesis::gen_key_to_store;

pub use store::AddressVpType;

pub use store::Store;

pub use self::store::ConfirmationResponse;

pub use self::store::ValidatorData;

pub use self::store::ValidatorKeys;

Modules

alias

Wallet address and key aliases.

fs

Implementation of wallet functionality depending on a standard filesystem

pre_genesis

Provides functionality for managing validator keys

store

Wallet Store information

test_utils

Structs

DatedKeypair

A keypair with a block height after which it was created

DerivationPath

A key derivation path

StoreSpendingKey

Extended spending key with Borsh serialization compatible with DatedSpendingKey. This is necessary to facilitate reading the old Store format.

Wallet

Represents a collection of keys and addresses while caching key decryptions

Enums

DecryptionError

DerivationPathError

FindKeyError

The error that is produced when a given key cannot be obtained

LoadStoreError

Errors of wallet loading and storing

StoredKeypair

A keypair stored in a wallet

Traits

WalletIo

Captures the interactive parts of the wallet’s functioning

WalletStorage

Captures the permanent storage parts of the wallet’s functioning

Functions

gen_secret_key

Generate a new secret key.

Type Aliases

DatedSpendingKey

Type alias for a spending key with a birthday.

DatedViewingKey

Type alias for a viewing key with a birthday.