namada_wallet
Struct Wallet

pub struct Wallet<U> { /* private fields */ }

Represents a collection of keys and addresses while caching key decryptions

Implementations

impl<U> Wallet<U>

pub fn new(utils: U, store: Store) -> Self

Create a new wallet from the given backing store and storage location

pub fn add_validator_data(&mut self, address: Address, keys: ValidatorKeys)

Add validator data to the store

pub fn get_validator_data(&self) -> Option<&ValidatorData>

Returns a reference to the validator data, if it exists.

pub fn get_validator_data_mut(&mut self) -> Option<&mut ValidatorData>

Returns a mut reference to the validator data, if it exists.

pub fn take_validator_data(&mut self) -> Option<ValidatorData>

Take the validator data, if it exists.

pub fn into_validator_data(self) -> Option<ValidatorData>

Returns the validator data, if it exists.

pub fn store(&self) -> &Store

Provide immutable access to the backing store

pub fn store_mut(&mut self) -> &mut Store

Provide mutable access to the backing store

pub fn extend_from_pre_genesis_validator( &mut self, validator_address: Address, validator_alias: Alias, other: ValidatorWallet, )

Extend this wallet from pre-genesis validator wallet.

pub fn get_addresses_with_vp_type( &self, vp_type: AddressVpType, ) -> HashSet<Address>

Gets all addresses given a vp_type

pub fn add_vp_type_to_address( &mut self, vp_type: AddressVpType, address: Address, )

Add a vp_type to a given address

pub fn tokens_with_aliases(&self) -> BTreeMap<String, Address>

Get addresses with tokens VP type keyed and ordered by their aliases.

pub fn find_address(&self, alias: impl AsRef<str>) -> Option<Cow<'_, Address>>

Find the stored address by an alias.

pub fn find_alias(&self, address: &Address) -> Option<&Alias>

Find an alias by the address if it’s in the wallet.

pub fn lookup_alias(&self, addr: &Address) -> String

Try to find an alias for a given address from the wallet. If not found, formats the address into a string.

pub fn lookup_ibc_token_alias(&self, ibc_denom: impl AsRef<str>) -> String

Try to find an alias of the base token in the given IBC denomination from the wallet. If not found, formats the IBC denomination into a string.

pub fn find_viewing_key( &self, alias: impl AsRef<str>, ) -> Result<&ExtendedViewingKey, FindKeyError>

Find the viewing key with the given alias in the wallet and return it

pub fn find_birthday(&self, alias: impl AsRef<str>) -> Option<&BlockHeight>

Find the birthday of the given alias

pub fn find_payment_addr( &self, alias: impl AsRef<str>, ) -> Option<&PaymentAddress>

Find the payment address with the given alias in the wallet and return it

pub fn find_alias_by_payment_addr( &self, payment_address: &PaymentAddress, ) -> Option<&Alias>

Find an alias by the payment address if it’s in the wallet.

pub fn get_secret_keys( &self, ) -> HashMap<String, (&StoredKeypair<SecretKey>, Option<&PublicKeyHash>)>

Get all known keys by their alias, paired with PKH, if known.

pub fn get_public_keys(&self) -> HashMap<String, PublicKey>

Get all known public keys by their alias.

pub fn get_addresses(&self) -> HashMap<String, Address>

Get all known addresses by their alias, paired with PKH, if known.

pub fn get_payment_addrs(&self) -> HashMap<String, PaymentAddress>

Get all known payment addresses by their alias

pub fn get_viewing_keys(&self) -> HashMap<String, ExtendedViewingKey>

Get all known viewing keys by their alias

pub fn get_spending_keys( &self, ) -> HashMap<String, &StoredKeypair<StoreSpendingKey>>

Get all known viewing keys by their alias

pub fn is_encrypted_secret_key(&self, alias: impl AsRef<str>) -> Option<bool>

Check if alias is an encrypted secret key

pub fn is_encrypted_spending_key(&self, alias: impl AsRef<str>) -> Option<bool>

Check if alias is an encrypted spending key

impl<U: WalletStorage> Wallet<U>

pub fn load(&mut self) -> Result<(), LoadStoreError>

Load a wallet from the store file.

pub fn save(&self) -> Result<(), LoadStoreError>

Save the wallet store to a file.

impl<U: WalletIo> Wallet<U>

pub fn derive_store_spending_key_from_mnemonic_code( &mut self, alias: String, alias_force: bool, birthday: Option<BlockHeight>, unsafe_pure_zip32: bool, derivation_path: DerivationPath, mnemonic_passphrase: Option<(Mnemonic, Zeroizing<String>)>, prompt_bip39_passphrase: bool, password: Option<Zeroizing<String>>, ) -> Option<(String, ExtendedSpendingKey)>

Restore a spending key from the user mnemonic code (read from stdin) using a given ZIP32 derivation path and insert it into the store with the provided alias, converted to lower case. The key is encrypted with the provided password. If no password provided, will prompt for password from stdin. Stores the key in decrypted key cache and returns the alias of the key and a reference-counting pointer to the key.

pub fn find_path_by_viewing_key( &self, vk: &ExtendedViewingKey, ) -> Result<DerivationPath, FindKeyError>

Find a derivation path by viewing key

pub fn derive_store_key_from_mnemonic_code( &mut self, scheme: SchemeType, alias: Option<String>, alias_force: bool, derivation_path: DerivationPath, mnemonic_passphrase: Option<(Mnemonic, Zeroizing<String>)>, prompt_bip39_passphrase: bool, password: Option<Zeroizing<String>>, ) -> Option<(String, SecretKey)>

Restore a keypair from the user mnemonic code (read from stdin) using a given BIP44 derivation path and derive an implicit address from its public part and insert them into the store with the provided alias, converted to lower case. If none provided, the alias will be the public key hash (in lowercase too). The key is encrypted with the provided password. If no password provided, will prompt for password from stdin. Stores the key in decrypted key cache and returns the alias of the key and a reference-counting pointer to the key.

Any usage of this function should be careful not expose a shielded key that may be derived via modified ZIP32 from this key (specifically when the derivation path and schema matching DerivationPath::modified_zip32 and no encryption password is being used).

pub fn gen_store_spending_key( &mut self, alias: String, birthday: Option<BlockHeight>, password: Option<Zeroizing<String>>, force_alias: bool, csprng: &mut (impl CryptoRng + RngCore), ) -> Option<(String, ExtendedSpendingKey)>

Generate a spending key similarly to how it’s done for keypairs

pub fn gen_store_secret_key( &mut self, scheme: SchemeType, alias: Option<String>, alias_force: bool, password: Option<Zeroizing<String>>, rng: &mut (impl CryptoRng + RngCore), ) -> Option<(String, SecretKey)>

Generate a new keypair, derive an implicit address from its public key and insert them into the store with the provided alias, converted to lower case. If none provided, the alias will be the public key hash (in lowercase too). If the alias already exists, optionally force overwrite the keypair for the alias. If no encryption password is provided, the keypair will be stored raw without encryption. Stores the key in decrypted key cache and returns the alias of the key and a reference-counting pointer to the key.

pub fn gen_hd_seed( passphrase: Option<Zeroizing<String>>, rng: &mut U::Rng, prompt_bip39_passphrase: bool, ) -> (Mnemonic, Seed)

Generate a BIP39 mnemonic code, and derive HD wallet seed from it using the given passphrase. If no passphrase is provided, optionally prompt for a passphrase.

pub fn derive_store_hd_secret_key( &mut self, scheme: SchemeType, alias: Option<String>, alias_force: bool, seed: Seed, derivation_path: DerivationPath, password: Option<Zeroizing<String>>, ) -> Option<(String, SecretKey)>

Derive a keypair from the given seed and path, derive an implicit address from this keypair, and insert them into the store with the provided alias, converted to lower case. If none provided, the alias will be the public key hash (in lowercase too). If the alias already exists, optionally force overwrite the keypair for the alias. If no encryption password is provided, the keypair will be stored raw without encryption. Stores the key in decrypted key cache and returns the alias of the key and the key itself.

Any usage of this function should be careful not expose a shielded key that may be derived via modified ZIP32 from this key (specifically when the derivation path and schema matching DerivationPath::modified_zip32 and no encryption password is being used).

pub fn derive_store_hd_spending_key( &mut self, alias: String, force_alias: bool, birthday: Option<BlockHeight>, seed: Seed, derivation_path: DerivationPath, password: Option<Zeroizing<String>>, ) -> Option<(String, ExtendedSpendingKey)>

Derive a masp shielded key from the given seed and path, and insert it into the store with the provided alias, converted to lower case. If the alias already exists, optionally force overwrite the key for the alias.

An optional birthday can be provided saying that this key was created after this blockheight.

If no encryption password is provided, the key will be stored raw without encryption.

Stores the key in decrypted key cache and returns the alias of the key and the key itself.

pub fn gen_disposable_signing_key( &mut self, rng: &mut (impl CryptoRng + RngCore), ) -> SecretKey

Generate a disposable signing key for fee payment and store it under the precomputed alias in the wallet. This is simply a wrapper around gen_key to manage the alias

pub fn find_secret_key( &mut self, alias_pkh_or_pk: impl AsRef<str>, password: Option<Zeroizing<String>>, ) -> Result<SecretKey, FindKeyError>

Find the stored key by an alias, a public key hash or a public key. If the key is encrypted and password not supplied, then password will be interactively prompted. Any keys that are decrypted are stored in and read from a cache to avoid prompting for password multiple times.

pub fn find_public_key_from_implicit_addr( &self, implicit_addr: &ImplicitAddress, ) -> Result<PublicKey, FindKeyError>

Find a public key in the wallet from the given implicit address.

pub fn find_public_key( &self, alias_or_pkh: impl AsRef<str>, ) -> Result<PublicKey, FindKeyError>

Find the public key by an alias or a public key hash.

pub fn find_spending_key( &mut self, alias: impl AsRef<str>, password: Option<Zeroizing<String>>, ) -> Result<ExtendedSpendingKey, FindKeyError>

Find the spending key with the given alias in the wallet and return it. If the spending key is encrypted but a password is not supplied, then it will be interactively prompted.

pub fn find_key_by_pk( &mut self, pk: &PublicKey, password: Option<Zeroizing<String>>, ) -> Result<SecretKey, FindKeyError>

Find the stored key by a public key. If the key is encrypted and password not supplied, then password will be interactively prompted for. Any keys that are decrypted are stored in and read from a cache to avoid prompting for password multiple times.

pub fn find_path_by_pkh( &self, pkh: &PublicKeyHash, ) -> Result<DerivationPath, FindKeyError>

Find a derivation path by public key hash

pub fn find_public_key_by_pkh( &self, pkh: &PublicKeyHash, ) -> Result<PublicKey, FindKeyError>

Find the public key by a public key hash. If the key is encrypted and password not supplied, then password will be interactively prompted for. Any keys that are decrypted are stored in and read from a cache to avoid prompting for password multiple times.

pub fn find_key_by_pkh( &mut self, pkh: &PublicKeyHash, password: Option<Zeroizing<String>>, ) -> Result<SecretKey, FindKeyError>

Find the stored key by a public key hash. If the key is encrypted and password is not supplied, then password will be interactively prompted for. Any keys that are decrypted are stored in and read from a cache to avoid prompting for password multiple times.

pub fn insert_address( &mut self, alias: impl AsRef<str>, address: Address, force_alias: bool, ) -> Option<String>

Add a new address with the given alias. If the alias is already used, will ask whether the existing alias should be replaced, a different alias is desired, or the alias creation should be cancelled. Return the chosen alias if the address has been added, otherwise return nothing.

pub fn insert_keypair( &mut self, alias: String, alias_force: bool, sk: SecretKey, password: Option<Zeroizing<String>>, address: Option<Address>, path: Option<DerivationPath>, ) -> Option<String>

Add a new keypair with the given alias. If the alias is already used, will ask whether the existing alias should be replaced, a different alias is desired, or the alias creation should be cancelled. Return the chosen alias if the keypair has been added, otherwise return nothing.

pub fn insert_public_key( &mut self, alias: String, pubkey: PublicKey, address: Option<Address>, path: Option<DerivationPath>, force_alias: bool, ) -> Option<String>

Insert a new public key with the given alias. If the alias is already used, then display a prompt for overwrite confirmation.

pub fn insert_viewing_key( &mut self, alias: String, view_key: ExtendedViewingKey, birthday: Option<BlockHeight>, force_alias: bool, path: Option<DerivationPath>, ) -> Option<String>

Insert a viewing key into the wallet under the given alias

pub fn insert_spending_key( &mut self, alias: String, force_alias: bool, spend_key: ExtendedSpendingKey, birthday: Option<BlockHeight>, password: Option<Zeroizing<String>>, path: Option<DerivationPath>, ) -> Option<String>

Insert a spending key into the wallet under the given alias

pub fn insert_payment_addr( &mut self, alias: String, payment_addr: PaymentAddress, force_alias: bool, ) -> Option<String>

Insert a payment address into the wallet under the given alias

pub fn extend(&mut self, wallet: Self)

Extend this wallet from another wallet (typically pre-genesis). Note that this method ignores store.validator_data if any.

pub fn remove_all_by_alias(&mut self, alias: String)

Remove keys and addresses associated with the given alias