namada_wallet
Type Alias DatedSpendingKey

pub type DatedSpendingKey = DatedKeypair<ExtendedSpendingKey>;

Type alias for a spending key with a birthday.

Aliased Type

struct DatedSpendingKey {
    pub key: ExtendedSpendingKey,
    pub birthday: BlockHeight,
}

Fields
key: ExtendedSpendingKey

The keypair itself

birthday: BlockHeight

A blockheight that precedes the creation of the keypair