namada_wallet
Type Alias DatedViewingKey

pub type DatedViewingKey = DatedKeypair<ExtendedViewingKey>;

Type alias for a viewing key with a birthday.

Aliased Type

struct DatedViewingKey {
    pub key: ExtendedViewingKey,
    pub birthday: BlockHeight,
}

Fields
key: ExtendedViewingKey

The keypair itself

birthday: BlockHeight

A blockheight that precedes the creation of the keypair