namada_wallet
Trait WalletIo

pub trait WalletIo: Sized + Clone {
    type Rng: RngCore;

    // Provided methods
    fn generate_mnemonic_code(
        mnemonic_type: MnemonicType,
        rng: &mut Self::Rng,
    ) -> Mnemonic { ... }
    fn read_password(
        _confirm: bool,
        _target_key: Option<&str>,
    ) -> Zeroizing<String> { ... }
    fn read_alias(_prompt_msg: &str) -> String { ... }
    fn read_mnemonic_code() -> Option<Mnemonic> { ... }
    fn read_mnemonic_passphrase(_confirm: bool) -> Zeroizing<String> { ... }
    fn show_overwrite_confirmation(
        _alias: &Alias,
        _alias_for: &str,
    ) -> ConfirmationResponse { ... }
}

Captures the interactive parts of the wallet’s functioning

Required Associated Types

type Rng: RngCore

Secure random number generator

Provided Methods

fn generate_mnemonic_code( mnemonic_type: MnemonicType, rng: &mut Self::Rng, ) -> Mnemonic

Generates a random mnemonic of the given mnemonic type.

fn read_password(_confirm: bool, _target_key: Option<&str>) -> Zeroizing<String>

Read the password for decryption from the file/env/stdin.

fn read_alias(_prompt_msg: &str) -> String

Read an alias from the file/env/stdin.

fn read_mnemonic_code() -> Option<Mnemonic>

Read mnemonic code from the file/env/stdin.

fn read_mnemonic_passphrase(_confirm: bool) -> Zeroizing<String>

Read a mnemonic code from the file/env/stdin.

fn show_overwrite_confirmation( _alias: &Alias, _alias_for: &str, ) -> ConfirmationResponse

The given alias has been selected but conflicts with another alias in the store. Offer the user to either replace existing mapping, alter the chosen alias to a name of their choice, or cancel the aliasing.

Object Safety

This trait is not object safe.